chore: Moved helpers inside test
