+++ test/playwright_put.unit.spec.ts
@@ -1,12 +1,11 @@
 import {
-  mockPage,
   mockChromium,
   mockApiContext,
   mockRequest,
   mockFs,
   mockPath,
-} from "../helpers/mocks";
-import { resetAllMocks } from "../helpers/mockSetup";
+  resetAllMocks,
+} from "./helpers";
 
 // Mock the playwright module
 jest.mock("playwright", () => {
